msft-docs-agent/requirements.txt
azure-ai-agentserver-agentframework
agent-framework
azure-identity
